#10 Authorization

+++ src/main/java/org/example/taskmanager/service/implement/ProfileServiceImpl.java
@@ -7,57 +7,78 @@ import org.example.taskmanager.api.request.profile.PutProfileRequest;
 import org.example.taskmanager.api.response.ProfileResponse;
 import org.example.taskmanager.entity.Profile;
 import org.example.taskmanager.entity.Task;
-import org.example.taskmanager.exception.ExpectedEntityNotFoundException;
 import org.example.taskmanager.repository.ProfileDAO;
 import org.example.taskmanager.repository.TaskCommentDAO;
 import org.example.taskmanager.repository.TaskDAO;
 import org.example.taskmanager.service.interfaces.ProfileService;
-import org.modelmapper.ModelMapper;
+import org.example.taskmanager.util.PasswordUtil;
+import org.example.taskmanager.util.ProfileAccessUtil;
+import org.example.taskmanager.util.ProfileUtil;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 import java.util.Collection;
 import java.util.UUID;
+import java.util.logging.Logger;
 
 @Service
 @RequiredArgsConstructor
 public class ProfileServiceImpl implements ProfileService {
 
+    private static final Logger LOGGER = Logger.getLogger(ProfileServiceImpl.class.getName());
+
     private final ProfileDAO profileDAO;
-    private final ModelMapper modelMapper;
+    private final ProfileUtil profileUtil;
     private final TaskDAO taskDAO;
     private final TaskCommentDAO taskCommentDAO;
+    private final ProfileAccessUtil profileAccessUtil;
+    private final PasswordUtil passwordUtil;
 
     @Override
     @Transactional
-    public ProfileResponse create(AddProfileRequest profile) {
+    public ProfileResponse create(AddProfileRequest request) {
+        //TODO add permission status to Profile
+        //TODO checkAdminPermission on FilterChain level
+
         var id = UUID.randomUUID();
-        profileDAO.save(convertToEntity(id, profile));
+        profileDAO.save(new Profile(
+                id,
+                request.getName(),
+                request.getEmail(),
+                passwordUtil.getEncodedPassword(request.getPassword()),
+                true
+        ));
         return getById(id);
     }
 
     @Override
     public ProfileResponse getById(UUID id) {
-        return convertToResponse(getEntity(id));
+        return profileUtil.tryToGetProfile(id);
     }
 
     @Override
     @Transactional
     public ProfileResponse update(UUID id, PutProfileRequest request) {
-        profileDAO.save(convertToEntity(id, request));
+        // check
+        profileAccessUtil.checkAuthorAuthorization(id);
+        // logic
+        profileDAO.save(profileUtil.convertToEntity(id, request));
         return getById(id);
     }
 
     @Override
     @Transactional
     public Boolean delete(UUID id) {
+        // check
+        profileAccessUtil.checkAuthorAuthorization(id);
+        // logic
         if (thereAreNoTasksWithProfileAsAuthor(id) && thereAreNoTaskCommentsFromProfile(id)) {
             if (!thereAreNoTasksWithProfileAsExecutor(id)) {
                 removeProfileFromTaskAsExecutor(id);
             }
             profileDAO.deleteById(id);
         } else {
-            var deactivatedProfile = getEntity(id);
+            var deactivatedProfile = profileUtil.getEntity(id);
             deactivatedProfile.setIsActive(false);
             profileDAO.save(deactivatedProfile);
         }
@@ -69,19 +90,20 @@ public class ProfileServiceImpl implements ProfileService {
     @Override
     public Collection<ProfileResponse> getAllWithPagination(Integer pageSize, Integer pageNumber) {
         return profileDAO.findAll(PageRequest.of(pageNumber, pageSize)).stream()
-                .map(this::convertToResponse)
+                .map(profileUtil::convertToResponse)
                 .toList();
 
     }
 
     private void removeProfileFromTaskAsExecutor(UUID profileId) {
         Collection<Task> tasks = taskDAO.findTaskByExecutorId(profileId);
-        tasks.stream().map(it -> {
+        tasks.stream()
+                .map(it -> {
                     var newTaskVer = it.clone();
                     newTaskVer.setExecutorId(null);
                     return newTaskVer;
                 })
-                .toList().forEach(taskDAO::save);
+                .forEach(taskDAO::save);
     }
 
     private Boolean thereAreNoTasksWithProfileAsAuthor(UUID profileId) {
@@ -94,22 +116,5 @@ public class ProfileServiceImpl implements ProfileService {
 
     private Boolean thereAreNoTaskCommentsFromProfile(UUID profileID) {
         return !taskCommentDAO.existsByAuthorId(profileID);
-    }
-
-    private Profile getEntity(UUID id) {
-        return profileDAO.findById(id).orElseThrow(
-                () -> new ExpectedEntityNotFoundException(String.format("Profile (%s) not found", id))
-        );
-    }
-
-    private <T> Profile convertToEntity(UUID id, T dto) {
-        var profile = modelMapper.map(dto, Profile.class);
-        if (profile.getId() == null) profile.setId(id);
-        if (profile.getIsActive() == null) profile.setIsActive(true);
-        return profile;
-    }
-
-    private ProfileResponse convertToResponse(Profile entity) {
-        return modelMapper.map(entity, ProfileResponse.class);
     }
 }
